02langchain-tutorial/requirements.txt
langchain>=0.1.0
python-dotenv>=1.0.0
openai>=1.0.0
colorama>=0.4.6
langchain-core>=0.1.0
langchain-community>=0.0.10 